Add lock/unlock functionality to tooltips
I needed a way of locking a tooltip open until a certain action was
performed.
To do so, I added a #lock and a #unlock method to the Tooltip API. Now I
can force a tooltip to remain open and unlock it when I want it to resume
normal operation. The lock/unlock methods do not invoke show/hide
automatically, giving more control to the end user to how the tooltip
functions.